+++ test/ERC721Collection.t.sol
@@ -1,48 +1,40 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.13;
+
 import {Test, console} from "forge-std/Test.sol";
 import {Drop} from "../src/ERC721Collection.sol";
 import {IERC721Collection} from "../src/IERC721Collection.sol";
 
-contract ERC721CollectionTest is Test{
+contract ERC721CollectionTest is Test {
     Drop public collection;
-    uint public singleMintCost = 110;
-    uint publicMintLimit = 2;
+    uint256 public singleMintCost = 110;
+    uint256 publicMintLimit = 2;
     address platform = address(234);
     address creator = address(123);
+
     function setUp() public {
         deal(address(123), 200);
-        collection = new Drop(
-            "Test Collection",
-            "TST",
-            100,
-            0,
-            100,
-            10,
-            100,
-            2,
-           creator,
-            "https://example.com/",
-            platform
-        );
+        collection =
+            new Drop("Test Collection", "TST", 100, 0, 100, 10, 100, 2, creator, "https://example.com/", platform);
     }
 
-    function _mintPublic(address _to, uint _amount, uint _value) internal {
-        uint previousBalance = collection.balanceOf(address(_to));
-        uint previousTotalMinted = collection.totalMinted();
-        uint previousCreatorBalance = creator.balance;
+    function _mintPublic(address _to, uint256 _amount, uint256 _value) internal {
+        uint256 previousBalance = collection.balanceOf(address(_to));
+        uint256 previousTotalMinted = collection.totalMinted();
+        uint256 previousCreatorBalance = creator.balance;
         console.log("Creator previous balance: ", previousCreatorBalance);
-        uint previousPlatformBalance = platform.balance;
+        uint256 previousPlatformBalance = platform.balance;
         console.log("Platform previous balance: ", previousPlatformBalance);
         vm.prank(_to);
-        if (_amount > publicMintLimit){
+        if (_amount > publicMintLimit) {
             vm.expectRevert(IERC721Collection.PhaseLimitExceeded.selector);
             collection.mintPublic{value: _value}(_amount, address(_to));
-        }
-        else{
+        } else {
             collection.mintPublic{value: _value}(_amount, address(_to));
-            uint creatorNewBalance = previousCreatorBalance + collection.computeShare(_amount, IERC721Collection.Payees.CREATOR);
-            uint platformNewBalance = previousPlatformBalance + collection.computeShare(_amount, IERC721Collection.Payees.PLATFORM);
+            uint256 creatorNewBalance =
+                previousCreatorBalance + collection.computeShare(_amount, IERC721Collection.Payees.CREATOR);
+            uint256 platformNewBalance =
+                previousPlatformBalance + collection.computeShare(_amount, IERC721Collection.Payees.PLATFORM);
             console.log("Creator new balance: ", creatorNewBalance);
             console.log("Platform new balance: ", platformNewBalance);
             assertEq(collection.balanceOf(address(_to)), previousBalance + _amount);
@@ -52,27 +44,27 @@ contract ERC721CollectionTest is Test{
         }
     }
 
-    function test_publicMint() public{
+    function test_publicMint() public {
         address minter = address(456);
         deal(minter, 350);
         _mintPublic(minter, 1, singleMintCost);
     }
 
-    function testFail_publicMintLimit() public{
+    function testFail_publicMintLimit() public {
         address minter = address(456);
         deal(minter, 350);
         _mintPublic(minter, 3, singleMintCost * 3);
     }
 
-    function test_mintMultiple() public{
+    function test_mintMultiple() public {
         address minter = address(456);
         deal(minter, 350);
         _mintPublic(minter, 2, singleMintCost * 2);
     }
 
-    function test_addPresale() public{
-        uint startTimeInSec = 0;
-        uint endTimeInSec = 100;
+    function test_addPresale() public {
+        uint256 startTimeInSec = 0;
+        uint256 endTimeInSec = 100;
         IERC721Collection.PresalePhaseIn memory phase = IERC721Collection.PresalePhaseIn({
             maxPerAddress: 2,
             name: "Test Phase",
@@ -91,9 +83,9 @@ contract ERC721CollectionTest is Test{
         assertEq(collection.getPresaleData()[0].endTime, block.timestamp + endTimeInSec);
     }
 
-    function testFail_presaleLimit() public{
-        uint startTimeInSec = 0;
-        uint endTimeInSec = 100;
+    function testFail_presaleLimit() public {
+        uint256 startTimeInSec = 0;
+        uint256 endTimeInSec = 100;
         IERC721Collection.PresalePhaseIn memory phase = IERC721Collection.PresalePhaseIn({
             maxPerAddress: 2,
             name: "Test Phase",
@@ -112,14 +104,14 @@ contract ERC721CollectionTest is Test{
         vm.stopPrank();
     }
 
-    function testAirdrop() public{
+    function testAirdrop() public {
         vm.prank(address(123));
         collection.airdrop(address(234), 3);
         assertEq(collection.balanceOf(address(234)), 3);
         assertEq(collection.totalMinted(), 3);
     }
 
-    function test_batchAirdrop() public{
+    function test_batchAirdrop() public {
         address[] memory addresses = new address[](3);
         addresses[0] = address(234);
         addresses[1] = address(345);
@@ -132,18 +124,18 @@ contract ERC721CollectionTest is Test{
         assertEq(collection.totalMinted(), 6);
     }
 
-    function test_ReduceSupply() public{
+    function test_ReduceSupply() public {
         vm.prank(address(123));
         collection.reduceSupply(50);
         assertEq(collection.maxSupply(), 50);
-    }   
+    }
 
-    function testFail_ReduceSupply() public{
+    function testFail_ReduceSupply() public {
         vm.prank(address(123));
         collection.reduceSupply(150);
     }
 
-    function testFail_ReduceSupply2() public{
+    function testFail_ReduceSupply2() public {
         vm.deal(address(789), 250);
         vm.prank(address(789));
         _mintPublic(address(789), 2, singleMintCost * 2);
